requirements.txt
pypng
Flask
pillow
pyqrcode
pyperclip
pytesseract